Tests/LibWeb/Text/expected/async-scrolling/scroll-tree-shape.txt

scroll node count: 2
scroll nodes: 2
has viewport node: true
viewport parent: 0
child nodes: 1
child has same document as viewport: true
child parent is viewport: true
